backend/requirements.txt
Flask==2.3.2
pylint==2.17.4
bandit==1.7.5
pyyaml==6.0
GitPython==3.1.31
requests==2.31.0
pytest==7.4.0
python-magic==0.4.27